Move Dummy to tests
We should only use it for tests, emphasize it by moving it out of the
horizons package. Adjust imports as needed.

I had to remove the set_trace import from tests/__init__.py, otherwise
an import of Dummy for the test setup would trigger imports from fife
before we could patch it.
It's not a big loss, set_trace is for development only and can still be
imported from tests.game.

Closes: #2515