Improve language selection in scenario menu
When switching between scenarios and both support the currently selected
language, the selection will be kept.

When switching to a scenario that doesn't support the currently selected
language, we try to determine the next best one by looking at the system
locale. When selecting another map, we once again first try to use the
language from the settings.

Also rewrite some helper functions to be stateless.

Fixes: #2451