Fix double subscription in gamelobby window
So the fix is pretty simple. When working with the windowing system,
it's important to be careful. In the lifetime of a window (that is, the
window manager knows about it), it'll will be:

  * opened exactly once
  * closed exactly once
  * shown/hidden multiple times

So it's a hude red flag when we subscribe to messages in `show`
(remember, can be called multiple times), but unsubscribe in `close`
(only called once).

Error:

    Traceback (most recent call last):
      ...
      File "/home/eoc/UH/unknown-horizons/horizons/gui/windows.py", line 309, in close
        self._windows[-1].show()
      File "/home/eoc/UH/unknown-horizons/horizons/gui/modules/multiplayermenu.py", line 420, in show
        NetworkInterface().subscribe("lobbygame_chat", self._on_chat_message)
      File "/home/eoc/UH/unknown-horizons/horizons/messaging/messagebus.py", line 111, in subscribe
        raise Exception("Callback %s already subscribed to %s" % (callback, type))
    Exception: Callback <bound method GameLobby._on_chat_message of <horizons.gui.modules.multiplayermenu.GameLobby object at 0x5a6a850>> already subscribed to lobby

Refs: #2152