Fix random SP map re-generation triggering
When adjusting the StepSlider for the first time to change some of the
maps parameters, nothing happend. Only the second adjustment resulted in
map re-generation. The problem was, that we triggered the map change
before reading the slider value into our local dict. Switching the order
around, the first change results in a new map. Also, the value is the
current on and not the last one.