Adapt loading of animations to current fife interface
Our engine fife changed its interface, so
`fife.Animation.createAnimation()` was removed in favour of using the
animationmanager. This animationmanager can hold on to already created
animations and return them. Thus, instead of always creating new
animations, we can just ask the manager for a reference to an already
created Animation instead.