Fix small `get_res_icon_path()` widget leak
`horizons.gui.util.get_res_icon_path()` does the same thing as as
`horizons.gui.widgets.imagebuton`: aks fife to create an `Icon()` to
see if it's a valid image path. But other than the `ImageButton`,
`get_res_icon_path()` did forget to call `hide()` on that `Icon`. So the
`Icon` stayed alive, even though unused.

We now call `hide()` on that `Icon`.