Some more mypy fixes
The lambda in productiontabs is pretty evil. mypy was complaining that
both `inventory` and `res` weren't defined. And it is correct, at the
time of the lambda defintion both aren't defined, only because the
lambda is used in a context (the for loop) were both are getting defined
it works. Which is a bit surprising, so I moved the lambda inside the
loop to make the closure over inventory and res more obvious.