Fix display of additional scenario information in SP menu
Not sure which change caused the issue, but when a scenario & language
was selected, `find_map_filename` of the ScenarioMapWidget tried to find
the associated yaml file to read the metadata from.

The second parameter, the language, was the name, e.g. 'English' instead
of 'en', and therefore the lookup in the filesystem failed.

Refs: #1678