Simplify horizons/i18n/objecttranslations
While the end result may be less clear in its purpose, I think it is
better than before. Before, one could assume that the data structure is
actually used in the game, but for objects, it is not! Not line gui
translations, where the code has to actually be valid python.

We probably don't need this file at all, we might as well just put it
there temporarily for updating the pot/po files and delete it
afterwards.

Also, the whole formatting logic around ROWINDENT is gone now, that's a
good thing I believe. The old code was difficult to see through.