Remove init_environment call when generating atlases
Passing False means that we don't want to setup fife, which is all that
function is doing anyway.

This import probably caused some problems when the git version of the
game is installed in the system. While run_uh.py has no problem
importing horizons (it is in sys.path after all), the other way around
doesn't work. When inside horizons code, we don't know where run_uh.py
actually is, and the current working directory might not be in sys.path.

See #2716