Fix #2717 - tooltips are now on correct side of cursor
Tooltips were always shown right of the cursor, even when the cursor was too close to the right edge of the screen.
This is now fixed: when the tooltip would be out of the screen it is shown left of the cursor instead.
This is done by postponing positioning of the tooltip until its size is known.