refactor building and tile origin; fix #661 and #1379 (mostly)
Buildings and ground tiles are now drawn with the center of their area as origin.
Image shifts are used for when the center of the building/tile does not match with the center of the image

facing location functions have been replaced with equivalent functions that do the same with rotation.

This is incomplete: drawing locations for units still have to be fixed