Show tooltips below buttons for more elements
PR #2781 fixed the tooltip positions for imagebuttons, but the problem also applies to other elements.
These include the resource bar items, and settings menu items.
This fix also sets the tooltip position for these elements.

The minimap is ignored, since that tooltip still moves.
StatusIcons for buildings are also ignored since it would be impossible to find the correct position.