minimap rotation always matches main view rotation (fixes #2783)
The rotation of the main view used to be stored in the savegame, but the minimap rotation would reset everytime a session is started,
Because of this the minimap could have a different rotation than the main view.
This update fixes that by making the minimap read the view rotation whenever a session is started or when the screen is rotated.

Also added some parameter documentation for MinimapTransform