key settings use fife's listener instead of fifechan's (fix #2156)
The listener for setting hotkeys used to use fifechan's keynames.
These don't completely match fife's keynames (for example for numpad keys)
This change makes the settings use the same listener that is used in game.

Also, this undoes #2792.
That PR was a workaround, but only gives trouble now the underlying problem is fixed.