setup.py: Replace deprecated `distro.linux_distribution` call with `distro.id`
As per https://distro.readthedocs.io/en/latest/#distro.id

Return the distro ID of the current distribution, as a machine-readable string.

For a number of OS distributions, the returned distro ID value is reliable, in the sense
that it is documented and that it does not change across releases of the distribution.

This package maintains the following reliable distro ID values:

Distro ID	Distribution
“ubuntu”	Ubuntu
“debian”	Debian
“rhel”		RedHat Enterprise Linux
“centos”	CentOS
“fedora”	Fedora
“sles”		SUSE Linux Enterprise Server
“opensuse”	openSUSE
“amazon”	Amazon Linux
“arch”		Arch Linux
“cloudlinux”	CloudLinux OS
“exherbo”	Exherbo Linux
“gentoo”	GenToo Linux
“ibm_powerkvm”	IBM PowerKVM
“kvmibm”	KVM for IBM z Systems
“linuxmint”	Linux Mint
“mageia”	Mageia
“mandriva”	Mandriva Linux
“parallels”	Parallels
“pidora”	Pidora
“raspbian”	Raspbian
“oracle”	Oracle Linux (and Oracle Enterprise Linux)
“scientific”	Scientific Linux
“slackware”	Slackware
“xenserver”	XenServer
“openbsd”	OpenBSD
“netbsd”	NetBSD
“freebsd”	FreeBSD